made changes based on PR landing page from results page
