+++ pubspec.yaml
@@ -35,10 +35,11 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  mockito: ^5.0.0
+  mocktail: ^1.0.4
   
 
 dev_dependencies:
+  mocktail: ^1.0.0
   flutter_test:
     sdk: flutter
