+++ test/results_page_test.dart
@@ -11,9 +11,9 @@ void main(){
       final expectedWidget = find.text("End of Results");
 
       await tester.dragUntilVisible(
-          expectedWidget, //widget that should be visible before it scrolls
-          find.byType(ListView), //scrolling through Listview
-          const Offset(0, 100) //0 is starting point, + num is scrolling down
+          expectedWidget, 
+          find.byType(ListView), 
+          const Offset(0, 100) 
     );
   });
 }
